refactor: reservationTime 분리에 따른 수정
