+++ src/main/java/roomescape/reservation/application/ReservationService.java
@@ -1,0 +1,34 @@
+package roomescape.reservation.application;
+
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import roomescape.reservation.domain.ReservationRepository;
+import roomescape.reservation.presentation.dto.ReservationRequest;
+import roomescape.reservation.presentation.dto.ReservationResponse;
+import roomescape.time.domain.ReservationTime;
+import roomescape.time.domain.ReservationTimeRepository;
+
+@Service
+@RequiredArgsConstructor
+public class ReservationService {
+
+    private final ReservationRepository reservationRepository;
+    private final ReservationTimeRepository timeRepository;
+
+    public List<ReservationResponse> getReservations() {
+        return reservationRepository.findAll()
+                .stream()
+                .map(ReservationResponse::from)
+                .toList();
+    }
+
+    public ReservationResponse addReservation(ReservationRequest request) {
+        ReservationTime time = timeRepository.findById(request.timeId());
+        return ReservationResponse.from(reservationRepository.save(ReservationRequest.toEntity(request, time)));
+    }
+
+    public void cancelReservation(Long id) {
+        reservationRepository.deleteById(id);
+    }
+}
